requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.9.0
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.2.4
django-axes==6.1.1
django-cors-headers==4.7.0
django-csp==3.7
django-environ==0.12.0
django-htmx==1.23.2
django-ratelimit==4.1.0
django-recaptcha==4.1.0
djangorestframework==3.16.1
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==25.0
passlib==1.7.4
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.20
python-pptx==1.0.2
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
sympy==1.13.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.6.0
XlsxWriter==3.2.0
